+++ lib/src/driver_api/index.dart
@@ -1,8 +1,5 @@
-export 'action.dart';
-export 'event.dart';
 export 'ui_controller.dart';
 export 'ui_driver.dart';
-export 'parser.dart';
 export 'http_meta.dart';
 export 'script_def.dart';
 export 'dependency.dart';