+++ lib/src/index.dart
@@ -4,3 +4,5 @@ export 'ui/index.dart';
 export 'driver_api/index.dart';
 export 'view_attributes/index.dart';
 export 'misc/index.dart';
+export 'animation_api/index.dart';
+export 'action_api/index.dart';