+++ lib/src/misc/index.dart
@@ -1,1 +1,3 @@
 export 'script_runner.dart';
+export 'logger.dart';
+export 'parser.dart';